+++ CHANGELOG.md
@@ -12,7 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Add `VlenCodec::with_index_location()`
 - Add `numcodecs.adler32` codec
 - Add `ChunkCacheTypePartialDecoder`, `ChunkCachePartialDecoderLru{Chunk,Size}Limit[ThreadLocal]`
-- Add `Array::storage()`
+- Add `Array::storage()` and `Array::with_storage()`
 - Add `Array<T>::[async_]readable()` where `T: [Async]ReadableWritableStorageTraits`
 
 ### Changed